\name{clust-class}
\Rdversion{1.1}
\docType{class}
\alias{clust-class}

\title{Class \code{"clust"}}
\description{
Vector where v[i] represents the most likely cluster of the data which sequence i was drawn from.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("clust", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{clust}:}{Object of class \code{"numeric"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "clust" in the signature.
}
\author{
Erik Gregory
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
#showClass("clust")
}
\keyword{classes}
