\docType{data}
\name{datas}
\alias{datas}
\title{Sample data set.}
\format{The format is: num [1:1700, 1:15] 1 15 1 15 15 12 10 4 1 15 ...}
\description{
  Simple synthetic data set containing 3 modal sequences in
  15! space, with some noise added.
}
\examples{
data(datas)
head(datas)
}
\keyword{datasets}

