\name{dists-class}
\Rdversion{1.1}
\docType{class}
\alias{dists-class}

\title{Class \code{"dists"}}
\description{
A matrix where M[i, j] represents the weighted kendall distance from sequence i to modal sequence j.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("dists", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{dists}:}{Object of class \code{"matrix"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "dists" in the signature.
}

\author{
Erik Gregory
}
%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
showClass("dists")
}
\keyword{classes}
