\name{info-class}
\Rdversion{1.1}
\docType{class}
\alias{info-class}

\title{Class \code{"info"}}
\description{
Matrix containing all of the information necessary to calculate the Kendall distance from one ranking to another.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("info", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{info}:}{Object of class \code{"matrix"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "info" in the signature.
}

\author{
Erik Gregory
}

\examples{
showClass("info")
}
\keyword{classes}
