\name{items-class}
\Rdversion{1.1}
\docType{class}
\alias{items-class}

\title{Class \code{"items"}}
\description{
The number of items ranked in the data
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("items", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{items}:}{Object of class \code{"integer"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "items" in the signature.
}

\author{
Erik Gregory
}
%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
showClass("items")
}
\keyword{classes}
