\name{likelihoodGMM}
\alias{likelihoodGMM}
\title{Calculate the likelihood of a set of data in a GMM}
\usage{
  likelihoodGMM(ps, thetas, ranks)
}
\arguments{
  \item{ps}{Proportion of the data which was drawn from
  each cluster.}

  \item{thetas}{matrix of theta parameters for each
  cluster}

  \item{ranks}{a list of objects of class 'ranking', as
  generated by RMallow, one for each cluster.}
}
\value{
  The log-likelihood of the parameters, given the data.
}
\description{
  Calculate the likelihood of a set of data in a GMM
}

