\name{log.like-class}
\Rdversion{1.1}
\docType{class}
\alias{log.like-class}

\title{Class \code{"log.like"}}
\description{
A vector representing the log-likelihood at each iteration of the EM algorithm
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("log.like", ...)}.

}
\section{Slots}{
  \describe{
    \item{\code{log.like}:}{Object of class \code{"numeric"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "log.like" in the signature.
}

\author{
Erik Gregory
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
showClass("log.like")
}
\keyword{classes}
