\name{modes-class}
\Rdversion{1.1}
\docType{class}
\alias{modes-class}

\title{Class \code{"modes"}}
\description{
A matrix M where M[i, j] is an interger representing the jth item in the ith modal sequence
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("modes", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{modes}:}{Object of class \code{"matrix"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "modes" in the signature.
}

\author{
Erik Gregory
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
showClass("modes")
}
\keyword{classes}
