\name{neighbors}
\alias{neighbors}
\title{Function to grab all permutations which are generated
by a transposition of adjacent items}
\usage{
  neighbors(sig)
}
\arguments{
  \item{sig}{a vector ranking to find the neighbors of}
}
\value{
  a matrix in which the rows are the neighbors of sig
}
\description{
  Function to grab all permutations which are generated by
  a transposition of adjacent items
}
\examples{
neighbors(c(1, 2, 3, 4))
}

