\name{props-class}
\Rdversion{1.1}
\docType{class}
\alias{props-class}

\title{Class \code{"props"}}
\description{
Vector where v[i] represents the estimated proportion of the data which was drawn from cluster i. 
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("props", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{props}:}{Object of class \code{"numeric"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "props" in the signature.
}

\author{
Erik Gregory
}
%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
showClass("props")
}
\keyword{classes}
