\name{pvals-class}
\Rdversion{1.1}
\docType{class}
\alias{pvals-class}

\title{Class \code{"pvals"}}
\description{
Matrix where M[i, j] represents the probability that ranker i belongs to cluster j.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("pvals", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{pvals}:}{Object of class \code{"matrix"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "pvals" in the signature.
}
\author{
Erik Gregory
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
showClass("pvals")
}
\keyword{classes}
