\name{rankers-class}
\Rdversion{1.1}
\docType{class}
\alias{rankers-class}

\title{Class \code{"rankers"}}
\description{
The number of rankers in the data set.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("rankers", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{rankers}:}{Object of class \code{"integer"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "rankers" in the signature.
}
\author{
Erik Gregory
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
showClass("rankers")
}
\keyword{classes}
