\name{ranks-class}
\Rdversion{1.1}
\docType{class}
\alias{ranks-class}

\title{Class \code{"ranks"}}
\description{
A matrix where M[i, j] is the ranking of item j by ranker i.  These are called direct rankings.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("ranks", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{ranks}:}{Object of class \code{"data.frame"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "ranks" in the signature.
}

\author{
Erik Gregory
}

\examples{
showClass("ranks")
}
\keyword{classes}
