\name{relative-class}
\Rdversion{1.1}
\docType{class}
\alias{relative-class}

\title{Class \code{"relative"}}
\description{
Matrix where M[i, j] represents the number of items M[i, j:n] ranked M[i, j] by ranker i.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("relative", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{relative}:}{Object of class \code{"matrix"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "relative" in the signature.
}

\author{
Erik Gregory
}

\examples{
showClass("relative")
}
\keyword{classes}
