\docType{package}
\name{RMallow-package}
\alias{RMallow}
\alias{RMallow-package}
\title{Fit Multi-modal Mallows' models to ranking data.}
\description{
  Fits the Mallows' model to ranking data.  Data can be
  partially or fully-ranked.
}
\details{
  \tabular{ll}{ Package: \tab RMallow\cr Type: \tab
  Package\cr Version: \tab 1.0\cr Date: \tab 2012-02-18\cr
  License: \tab GPL (>= 2) }
}
\author{
  Erik Gregory Maintainer: <egregory2007@yahoo.com>
}
\references{
  "Mixtures of distance-based models for ranking data".
  Thomas Brendan Murphy & Donal Martin. 1 April 2002.
  Computational Statistics & Data Analysis 41 (2003)
  645-655.

  "Estimating a Population Distribution of Sequences of k
  Items from Cross- Sectional Data". Laurel A. Smith
  (Beckett) and Denis A. Evans. Journal of the Royal
  Statistical Society.  Series C (Applied Statistics). Vol.
  40, No. 1 (1991), pp.31-42.  Blackwell Publishing for the
  Royal Statistical Society. Accessed 16/08/2010.
  http://www.jstor.org/stable/2347903 .

  "A Non-iterative procedure for maximum likelihood
  estimation of the parameters of Mallows' Model Based on
  Partial Rankings".  Laura Adkins and Michael Flinger.
  Communication in Statistics - Theory and Methods, 27:9,
  2199-2220. 1998, Marchel Dekker, Inc.
  http://dx.doi.org/10.1080/03610929808832223 .
}
\keyword{ranking}

