\name{searchPi}
\alias{searchPi}
\title{Find the modal sequence using a \code{Q} matrix.}
\usage{
  searchPi(Qs, thetas)
}
\arguments{
  \item{Qs}{a transitions matrix of a data set}

  \item{thetas}{a matrix of weights on the metric}
}
\value{
  the best possible modal sequence for the cluster
}
\description{
  Find the modal sequence using a \code{Q} matrix.
}

