\name{theta-class}
\Rdversion{1.1}
\docType{class}
\alias{theta-class}

\title{Class \code{"theta"}}
\description{
The weights on the weighted Kendall metric.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("theta", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{theta}:}{Object of class \code{"matrix"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "theta" in the signature.
}
\author{
Erik Gregory
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
showClass("theta")
}
\keyword{classes}
