\name{ties-class}
\Rdversion{1.1}
\docType{class}
\alias{ties-class}

\title{Class \code{"ties"}}
\description{
A matrix where M[i, j] represents the number of items which ranker i ranked as j.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("ties", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{ties}:}{Object of class \code{"matrix"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "ties" in the signature.
}

\author{
Erik Gregory
}

\examples{
showClass("ties")
}
\keyword{classes}
