\name{type-class}
\Rdversion{1.1}
\docType{class}
\alias{type-class}

\title{Class \code{"type"}}
\description{
Either 'full' or 'partial', represents the type of rankings present in the data.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("type", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{type}:}{Object of class \code{"character"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "type" in the signature.
}

\author{
Erik Gregory
}
%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
showClass("type")
}
\keyword{classes}
