\name{UpdateP}
\alias{UpdateP}
\title{Update Proportion in each cluster.}
\usage{
  UpdateP(z)
}
\arguments{
  \item{z}{Probabilities that each sequence is in each
  cluster.}
}
\value{
  Proportion of data in each cluster.
}
\description{
  Updates the proportion of data assigned to each cluster.
}
\author{
  Erik Gregory
}
\references{
  "Mixtures of distance-based models for ranking data".
  Thomas Brendan Murphy & Donal Martin. 1 April 2002.
  Computational Statistics & Data Analysis 41 (2003)
  645-655.
}
\keyword{proportion}

